:recycle: Destructurate again
